arch/arm64: Add lazy preempt support
arm64 is missing support for PREEMPT_RT. The main feature which is
lacking is support for lazy preemption. The arch-specific entry code,
thread information structure definitions, and associated data tables
have to be extended to provide this support. Then the Kconfig file has
to be extended to indicate the support is available, and also to
indicate that support for full RT preemption is now available.

Change-Id: I18bcc6add8cc369661aaf3d0e97c73932b0ed82f
Signed-off-by: Anders Roxell <anders.roxell@linaro.org>
Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>