drivers/input: aw8697_haptic: allow modifying gain levels
the driver was supposed to modify the gain level when it is setup
using `aw8697_haptic_set_gain` but it fails to do so because the
driver does not allow chaning gain levels via the i2c nodes.
Instead use the work based node to set the haptic gain level.

Change-Id: I53ad49960a5fe350fc92222e6a9db85283919839
Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>